fix: IntroMusic now reduce its volume when in configuration.
Before this commit the volume stood the same for all the submenus, now when in a submenu the IntroMusic is reduced.

+++ SmartPiano_client/src/Model/AudioPlayer.java
@@ -60,7 +60,7 @@ public class AudioPlayer extends Thread {
         }
     }
 
-    public void setVolume(int db){
+    public void setVolume(float db){
         volume.setValue(db);
     }
 }
